First request

+++ src/main/java/br/com/consistency/Consistency/web/rest/UserResource.java
@@ -6,6 +6,7 @@ import br.com.consistency.Consistency.model.User;
 import br.com.consistency.Consistency.service.RequestService;
 import br.com.consistency.Consistency.service.UserService;
 import br.com.consistency.Consistency.util.HeaderUtil;
+import br.com.consistency.Consistency.util.Util;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,6 +18,7 @@ import javax.validation.Valid;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.List;
+import java.util.concurrent.ThreadLocalRandom;
 
 import static java.lang.String.format;
 
@@ -38,15 +40,20 @@ public class UserResource {
     @CrossOrigin
     @GetMapping("/user/get/{id}/{from}")
     public ResponseEntity<User> findById(@PathVariable Long id, @PathVariable Integer from) {
-        User user = userService.findById(id);
+        User user = null;
+        if (Config.enabled) {
+            user = userService.findById(id);
+        }
         return new ResponseEntity<User>(user, HttpStatus.OK);
     }
 
     @CrossOrigin
     @GetMapping("/user/{from}")
     public ResponseEntity<List<User>> findAll(@PathVariable Integer from) {
-//        System.out.println(new RequestService().request("8080", "users", "users"));
-        List<User> users = userService.findAll();
+        List<User> users = null;
+        if (Config.enabled) {
+            users = userService.findAll();
+        }
         return new ResponseEntity<List<User>>(users, HttpStatus.OK);
     }
 
@@ -55,7 +62,30 @@ public class UserResource {
     public ResponseEntity<User> createUser(@PathVariable String nome, @PathVariable Integer from) {
         String strLog = format(REST_REQUEST_FOR_SAVE, nome);
         log.debug(strLog);
-        User result = userService.save(nome);
+        User result = null;
+        if (Config.primary) {
+            if (Config.enabled) {
+                result = userService.save(nome);
+                RequestService.broardCastNewUser(nome);
+            } else {
+                RequestService.updatePrimary();
+            }
+        } else if (Config.enabled) {
+            RequestService.request("8080", "Add new User", "user/addFromPrimary/" + nome);
+        }
+        return new ResponseEntity<User>(result, HttpStatus.OK);
+    }
+
+    @CrossOrigin
+    @GetMapping("/user/addFromPrimary/{nome}/{from}")
+    public ResponseEntity<User> createUserPrimary(@PathVariable String nome, @PathVariable Integer from) {
+        String strLog = format(REST_REQUEST_FOR_SAVE, nome);
+        log.debug(strLog);
+        User result = null;
+        if (Config.enabled) {
+            Util.delay();
+            result = userService.save(nome);
+        }
         return new ResponseEntity<User>(result, HttpStatus.OK);
     }
 
@@ -64,7 +94,23 @@ public class UserResource {
     public ResponseEntity<Void> removeUser(@PathVariable Long id, @PathVariable Integer from) {
         String strLog = format(REST_REQUEST_FOR_DELETE, id);
         log.debug(strLog);
-        userService.delete(id);
+        if (Config.enabled) {
+            userService.delete(id);
+        }
+        return new ResponseEntity<Void>(HttpStatus.OK);
+    }
+
+    @CrossOrigin
+    @GetMapping("/user/disable/{from}")
+    public ResponseEntity<Void> disable(@PathVariable Integer from) {
+        Config.enabled = false;
+        return new ResponseEntity<Void>(HttpStatus.OK);
+    }
+
+    @CrossOrigin
+    @GetMapping("/user/enable/{from}")
+    public ResponseEntity<Void> enable(@PathVariable Integer from) {
+        Config.enabled = true;
         return new ResponseEntity<Void>(HttpStatus.OK);
     }
 }
